Added pylib directory for holding python modules I'd like to keep around.
adhoc.py contains the AdHoc class, which subclasses dict and allows keys
to be accessed using member notation as well as subscript notation. So
d.key yields the same value as k['key']. Makes things nice, so long as
all keys are legal identifiers. This is particularly nice when working
with database rows.

exiftool.py contains a python wrapper around Phil Harvey's exiftool
command, which is just called with subprocess.Popen(). I needed this
for chronorenme.