Added ability to read Excel spreadsheets directly. I'm using John Machin's excellent xlrd module to do this. http://pypi.python.org/pypi/xlrd
To feed an XLS file to this program, you have to name the file on the
command line (because it's the file extension that tell cutcsv to read
the file as an Excel spreadsheet. Only the fist worksheet in the
spreadsheet is read.

TODO: Add a --worksheet=N option to tell cutcsv to read from a worksteet
other than the first one.