json2csv is new and pretty-json (formerly json) is better.
json2csv reads JSON data starting at a given point in the data's
hierarchy and then writes it to standard output as CSV.

pretty-json (formerly json) supports an --indent N option, which
defaults to 2. It also outputs valid JSON in human readable form. It was
humanly readable before, but I was using pprint for output, which messed
up the JSON-ness of the data.