The ansi script and ansi.py module send colors to the terminal.
I'm starting to use ANSI escape sequences in enough code that it's
become worth while to modularlize that capability. Now that this has
been done, the next step is to update the mark script to use this new
ansi.py module rather than its own internal code for doing the same
thing.

Now that ANSI-fying code is become simpler and more "standard," I hope
to put it to use in lots of things. (As always, we'll see.)