Improved tread's format discrimination.
The thread script's date format was accepting 'mm/dd/yy' as if it were
'yy/mm/dd', and this was causing trouble (e.g. when interpreting
'mm/dd/yy' from Redhat's klist output).

I've added 'yyyy/mm/dd' as a pattern that insists on a 4-digit year and
'mm/dd/yy' that insists on a 2-digit year. This solves only the
immediate problem. Also, the parser now returns all matching
possibilities. My hope is that downstream logic can choose the best
match using some sort of heuristic. For the time being, it simply
complains and terminates when the few or more than one pattern matches a
given time string.

There's a new --tzlist option that outputs the time zones this script
understands.

TODO: Add architecture for choosing the best of multiple matches,
possibly allowing the calling code to give hints or express preferences.