Improved grep.py module
NEW: find_all keyword argument to Grep.__call__() tells it to return
even the strings that did not match any pattern.

FIXED: Grep.__call__() was returning the list of all result rather than
the list of only the matching results.

This module is very nearly ready to become the common foundation for the
existing mark and pygrep scripts. It will also be integral to the
upcoming "not" script that filters filenames.