grep.py gets bug fixes and much more versatility.
I've removed addict support entirely, in favor of using anonymous
objects for the same purpose. So Match.groupDict now returns either
match.groupdict() or None (if there was no match).

Grep's constructor now accepts all the same keyword arguments that its
__call__() method does. These become __call__()'s defaults, which makes
__call__() potentially simpler to call.

Also, there's a new "value" keyword argument which, if given, is a
function that accecpts the value of one item in the list of items to be
searched and returns the string value to use in the search on that item.
This comes in very handy if you need to grep through non-string values.
The only requirement is that the function accepts a single argument and
returns a string value.