Added --enumerate option to chronorename.
There was a bug in chronorename failed to increment the numeric suffix
beyond its initial value when needed. I've corrected this.

And since I found this bug while trying to simply number some files
(with no time embedded in the filenames), I've gone ahead and
implemented a --enumerate option that makes this easier from the command
line.