cutcsv now uses the the system line terminator
I was neglecting to specify the line terminator string when constructing
a CSV dialect, so that attribute was defaulting to '\r\n'. I've
"hardcoded" it for now to be the native line terminator for the system
the command is being run on.

TODO: This really SHOULD be something that can be set using the --reader
and --writer command line options. Since this is the only
multi-character dialect attribute, it will need to go at the end of the
dialect_spec value.