Downgraded cutcsv for use with Python ver <= 2.5
The "except Exception as e" syntax wasn't supported until Python 2.6,
and I need code to run on ... brace yourself .. Python 2.3 (ug). So I've
downgraded the code to use the old "except Exception,e" syntax (which
will crash if we try to run this under Python version >= 3.0, but how
likely is THAT?).