encode64 behaves better with std(in|out) now
encode64 (and therefore the symlinked decode64 as well) was reopening
/dev/stdout in order to set the mode to "ab". This worked everywhere but
on Solaris. I was committing a similar impropriety with stdin as well.
I've converted this code to use os.fdopen(file.fileno(),...) instead.
This makes Solaris much happier ... and it's what I should have done in
the first place.

Since the base64 command is less than universal, I've also explained to
encode64 that it should default to "encode" mode if it was called by
either "encode64" or "base64". (We apparently don't install the base64
command by default on our Solaris machines.)