New --log-parser option for pygrep
pygrep has a new --log-parser option that matches the RE on the command
line only against the message portion of each log line. All components
of the line are available as named (or numbered if you'd prefer) groups.
This greatly reduces the challenge of using pygrep for parsing and
formatting matching log data.