Adding semver.py (SemanticVersion class)
This is the start of a SemanticVersion class that understands semantic
version syntax and supports handling files with this type of naming
convention. I'm thinking about using this convention on all deployments
(which I haven't figured out yet), making SemanticVersion the basis of a
consistent --version command line option, and being able to handle
different versions of commands and Python modules that aren't
necessarily interface-compatible with previous versions of themselves.

See https://semver.org for more details.