Added simple relative day syntax to strptime
[(this time | end of)] (now | today | tomorrow | yesterday)

I know that's far from complete, but here's how it works at this point.
Case is completely insignificant for the entire phrase.

o "Now" yields the currtime date and time.

o "Today" is just like it, but the time of day is set to midnight (the
  beginning of today).

o "Tomorrow" and "yesterday" are like "today" offset plus or minus one
  day.

o By default "today", "tomorrow", and "yesterday" will yield 00:00:00
  (midnight) on the day in question. You can suppress this behavior by
  beginning the phrase with "this time" (as in "this time tomorrow". Or
  you could go the other way and request the last instant of the day in
  question by beginning the phrase with "end of" (as in "end of today"),
  which would yeld today's date with a time of 23:59:59.999999.