Bug fix: Month names after July were wrong
Because of poorly formatted string literal (containing month names),
August was missing, and subsequent months had the wrong number (for
instance, 8 was being returned for September). This is now fixed.

I'm also in the middle of adding docstrings for unit testing. And this
has be adding a some debug code I've been thinking about for a while.
(Look for DebugChannel in an upcoming commit.)