Added new test mode
I was debugging a problem that turned out to be the result of a terminal
feature rather than any issue with the code. But along the way I added
the ability to pass a color specification on the command line. So

    python ansi.py green on blue

outputs the text "green on blue" in green text on a blue background ...
depending on how your terminal behaves. :-)