Updating the csv command
LOTS of changes to csv ...

Usage:
  Form 1: csv [OPTIONS] [column-1 ... column-n]
    Input MUST be formatted as parsable CSV.

  Form 2: csv --excel [--worksheet=WORKSHEET] [OPTIONS] [column-1 ... column-n]
    Input MUST be an Excel spreadsheet. The first worksheet will be used unless named.

  Form 3: csv --shell [OPTIONS] [column-1 ... column-n]
    Input MUST be lines of shell-escaped and -quoted values.

This command reads data from standard input in either CSV (by default),
Excel, or shell format and writes it to standard output as properly
formatted CSV. The CSV dialects used for reading and writing need not be
the same and can be specified using the --reader (in the case of Form 1)
and --writer options.

Regardless of which usage form is used, command line arguments can be
given to supply a row of data that will be appear as the first line of
output. Typically, this is used to output column headings, (e.g.: 'First
name' Last\ name email account zipcode), but it could just as easily be
a row of data. Just in case it's not obvious, any data given on the
command line in this way is not CSV-formatted. It is parsed by your
shell, so use proper escapes and quoting to distinguish one argument
from another.

See "csv -h" for more.