CLI changes and format updates for csv command
I replaced --csv, --excel, and --shell with --infmt and added --outfmt,
which supports csv, shell, and table output formats. I haven't finished
debugging, but this is pretty close.

Usage: Usage:
  Form 1: csv [OPTIONS] [column-1 ... column-n]
    Input MUST be formatted as parsable CSV.

  Form 2: csv --infmt=excel [--worksheet=WORKSHEET] [OPTIONS] [column-1 ... column-n]
    Input MUST be an Excel spreadsheet. The first worksheet will be used unless named.

  Form 3: csv --infmt=shell [OPTIONS] [column-1 ... column-n]
    Input MUST be lines of shell-escaped and -quoted values.

This command reads data from standard input in either CSV (by default),
Excel, or shell format and writes it to standard output as properly
formatted CSV. The CSV dialects used for reading and writing need not be
the same and can be specified using the --reader (in the case of Form 1)
and --writer options.

Regardless of which usage form is used, command line arguments can be
given to supply a row of data that will be appear as the first line of
output. Typically, this is used to output column headings, (e.g.: 'First
name' Last\ name email account zipcode), but it could just as easily be
a row of data. Just in case it's not obvious, any data given on the
command line in this way is not CSV-formatted. It is parsed by your
shell, so use proper escapes and quoting to distinguish one argument
from another.

The output format defaults to CSV (--outfmt=csv), but it can also be set
to shell or table. If shell, each line of output is a list environment
variable assignments separated by semicolons. This REQUIRES that the
first line of input data consists of column headings that are valid when
used to name environment variables. If --outfmt=table is used, the first
row is also assumed to be column headings, but they can be any crazy
thing you want. (Don't take that as a challenge.)

Options:
  -h, --help
        show this help message and exit
  --reader=DIALECT
        Specify the CSV reader's dialect as
        SEP[Q[QSTYLE[DQUOTE[ESC[SKIPWS[EOL]]]]]]. This is a single
        string of up to five characters:
            SEP: Field separator character. (default: ,)
            Q: Quote character. (default: ")
            QSTYLE: Quoting style. One of 'a' (all), 'm' (minimal, the
        default), 'N' (non-numeric), or 'n' (none).
            DQUOTE: Represent a literal quote as two consecutive quotes.
        Either 't' (the default) or 'f'.
            ESC: The escape charater, which makes the next character a
        literal. (default: \)
            SKIPWS: Skip whitespace immediately following a field
        separator. Either 't' or 'f' (the default).
            EOL: The string that terminates every csv row. This is the
        only part of the dialect specification than can be more than one
        character in length. (default: '\n')
  --infmt=INFMT
        Specify either csv, excel, or shell as the input format. See the
        usage and description above for details. (default: csv)
  --outfmt=OUTFMT
        Specify either csv, shell, or table as the output format. See
        the usage and description above for details. (default: csv)
  --writer=DIALECT
        Specify the CSV writer's dialect, using the same syntax as for
        --reader. (default: ',"mt\\f\n')
  --worksheet=NAME_or_NUMBER
        Give the name or number (starting with 0) of the worksheet to
        read if --excel was used. If not given, the first worksheet will
        be read.
  --test
        Run internal tests (for debugging purposes only).