Let cutcsv exclude comments in CSV input
I've added a COMPLETELY NON-STANDARD feature to the CSV dialect: the
comment character specification. I think it's handy to be able to
comment lines out of a CSV file, so while this might be harrissy to
some, I think there's enough real-world, down-in-the-trenches value in
this feature to justify its non-conformity.

If this feature causes problems because my default comment character (#)
is causing problmes for your input data, try this:

    --reader ',"mt\f'

This leaves the comment character out of the reader's CSV dialect,
turning cutcsv's the ability to recornize and ignore CSV comments.