Don't let my copy of argparse obscure the real one
So that my borrowed (stolen? borrowed!) version of argparse.py from
Python 2.7 doesn't obscure any existing, native argparse module (which
only started appearing sometime during the 2.6 days), I've renamed it to
argparse27.py so that it can be imported using

try:
  import argparse
except:
  import argparse27 as argparse

if it's needed. This lets me use argparse under versions of Python as
far back as 2.4 (which is as far as I need to go).