Add a Table class for handling tabular data
The classes Table and Table.Row are defined here. Both are (completely
different) subclasses of Python's standard list. Table.Row instances are
tied to a table, which let them enforce the rule that all rows must have
the same number of columns. A Table instance is a list of rows that
knows how to enforce rules of and perform operations on tabular
(rectangular) data.

This code raises Table.Error exceptions when it abused.

THIS IS JUST AN INITIAL CHECK-IN. THERE ARE MORE FEATURES TO ADD AND
LOTS OF TESTING TO DO.