Simplify using the Mac's "say" command
This script reads aloud whatever it finds on standard input. The voice
used is the name of the script (alex, alice, fred, whatever). Use
symlinks to add or remove a voice. If you try to use a voice that the
"say" command can't handle, the script will show you the full list of
choices.

If you provide a filename on the command line, this script will "speak"
standard input into that mp3 file. A .mp3 file extension will be
provided if you leave that off.