Little improvements to loggy
get_log_level_by_name() would report an unrecognized log level in its
upper-case form. It now reports the same value it was passed.

Updated get_logger()'s doc string.

git_logger() was taking the basename of argv[0] and then removing
everything beginning with the first '.' character. It now removes only
the last file extension from the program name.

This still needs more testing, so this check-in is just to get it to a
system where testing is a little easier.