Improve make.py's isnewer(src,dest) function
isnewer(src,dst,bias=-0.001) is the new signature of this function. I've
added the bias argument, the number of seconds to add to the source
time, in order to bias the comparison with the time of the destination
file. It defaults to -0.001 seconds (-1 miliseconds) to avoid needless
file copies on some operating systems. (I'm looking at you, Mac OS.)