Improve json command
I've been working on some kind of system of managing json data. I want
to be able to basically treat it like a hierarchical data structure,
both addressing and searching it. This intermediate commit is a step in
that direction. Try out the "-O traverse" option. (This would be so much
easier in Python three with its "yield from" syntax.