Improve progress command with helpful return code
The prgress command now returns 0 if its position is less than its
limit. This lets you do things like this:

    i=0;while ./progress $i 10; do sleep 1; i=$((i+1)); done; echo

That's a contrived example, but it shows how the progress command's
return value can be used to control the logic surrounding that command.
More practically, here's how it might be used to monitor an rsync
command running in some parallel process.

    while progress $(du -h . | sed 's/G.*//') 79; do sleep 5; done; echo

It looks like I should also add an option to follow the progress
command's "last" rendering with a newline.