Add --quote-numbers to json command
Normally, numeric data is serialized to JSON format without quotes. The
new --quote-numbers option forces them to be quoted.

Actually, it occurs to me, the way I've implemented this was to prevent
already-quoted numeric data to be converted to numeric format when it
was read. I suppose I should re-implement this feature in the
BetterJsonEncoder class instead.