Finish ldif2csv command
This initial version of ldif2csv is funcitonal enough to begin using.
We'll see where it goes. It outputs the usage message below if given the
(hidden) --help command or if it can't find its input file on standard
input or named on the command line.

Usage: ldif2csv [optons] [infile]

Parse LDIF data from infile or standard input, and reformat it as CSV
to standard output (by default).

Options:
  -c COLUMNS, --columns=COLUMNS
                        Comma- and/or space-separated list of case-
                        sensitive attributes to include in CSV output.
                        If this option is not used, all attributes
                        found in the LDIF data will become columns and
                        will arranged alphabetically in the CSV
                        output. Use -h begin output with a row of
                        attribute names.
  --flatten             Flatten multi-valued attributes into as many
                        CSV rows as needed.
  --no-dn               Exclude DN from the output columns.
  -h                    Include a headings row. This is strongly
                        encouraged if --columns is not used.