Fix handy.first_match() function
I've removed the "flags=0" parameter from the first_match() function
because it turns out that RE flags are set when you compile an RE and
cannot be altered in a call to pattern.match() or pattern.search(). So
there's no point in passing such flags into first_match().

And now that Python and I are on the same page about this, the doctests
are all working.