Add CaselessString and CaselessDict classes
CaselessString subclasses unicode and defines case-insensitive
comparison and hashing operations for string values. Comparing
CaselessStrings with other strings or with other CaselessStrings is
always case-insensitive.

CaselessDict subclasses dict and uses CaselessString for key values when
given strings of any type to use for keys. The result is that d['a'] and
d['A'] refer to the same item.