Rename ldif2csv to ldif and implement subcommands
The ldif2csv command was about to spill the banks of its original
purpose, so I've renamed it to simply "ldif" and created at least the
scaffolding for the "csv", "diff", and "json" subcommands.

"ldif csv" will do what ldif2csv used to.

"ldif diff" will diff two LDIF inputs and output LDIF suitable for
feeding to ldapmodify.

"ldif json" will read LDIF and write it out as JSON.

There might be more, but I'm biting off quite enough as it is.