Fix warnings of bad services
Using --warn in the presence of "Name or service not known" services was
crashing certmon. This is no longer the case, such hosts are placed at
the top of the warning list.

I've also added a --timeout option because I don't like waiting a full
minute before a socket connection times out. (5 seconds is usually
plenty.)

I've also implemented persistent configuration in ~/.certmon.json if it
exists. Right now it's just storing a list of services (so I don't have
to keep them in a kludgy CSV file anymore), but there's also a "config"
section.

TODO: The next step should be to give certmon subcommnands. Maybe just
"config" for managing config data and "test" for actually testing certs.