Finish debugging versioning.py
I've renamed the DtpVersion class to RankedVersion and made the ranks
something the caller can set for himself. The next step (inviting
further debugging) is to figure out a good way to read and write version
data in source files.

---

The Version class is for storing, updating, and comparing semantic
version data. This is a very simple implementation, but it does allow
version comparisons to be numeric, so 1.5 < 1.10.

The meta value defaults to an empty string, but there's a fair amount of
versatility there too. Any numeric suffix is treated numerically so that
1.2.3.beta-5 < 1.2.3.beta-10, but the string portion is still compared
alphabetically, so 1.2.3.test-5 > 1.2.3.prod-5.

There's also the problem that 1.2.3 < 1.2.3.prod, but see the
RankedVersion for a solution to that.

---

RankedVersion is just like Version, except that it recognizes the rank
values "dev", "test", and "prod" (by default) in its meta component. For
purposes of comparing RankedVersion objects, each value is greater than
the one before it (in the order shown above).

A meta value of "" is greater than "prod", but any other non-ranked
value is less than "dev" and compares any with other non-ranked value
alphabetically.

Whatever the meta value is, any ranked value may be followed by an
arbitrary string that only participates in comparison operations if the
ranked part of each RankedVersion's meta value is the same. This may be
suffixed with a numeric value, which if present, will participate in
comparison operation according to its numeric value. So "1.2.3.prod.11"
is greater than "1.2.3.prod.2", and "1.2.3.beta-dog" is greater than
"1.2.3.beta-chimp", but remember that "1.2.3.test-anything" is greater
than "1.2.3.dev-something" That's what ranking is all about.