Replace pluralize() with decline() in handy.py
"Declension" is the linguist's word for what this function does, so I've
renamed it to decline(), just because it's nice for function names to be
verbs.

The delcine() function is way smarter than pluralize() was. It can
figure out the plural form of most english nouns. But it's still pretty
dumb in that it never heard of verbs. It treats everything as a noun.

The fmt argument is different, too. It's now a map-based format string,
with "count" and "word" being the only two map keys available to it. See
the usage text below.

Side note: The certmon script was using pluralize(), so I've updated
that to use decline() instead.

--------

decline(root, n, suffixes=None, fmt=None)
  This is all about performing numeric declension in English, though
  it's versatile enough to exceed that charter just a bit. For example,
  decline('box',2) returns 'boxes'.

  Arguments:
    root     - The root of the word whose proper declension is to be
               computed.
    n        - The number of the item in question.
    suffixes - An optional 2-element sequence of suffixes to be appended
               to root, depending on n's value. The sequence must be in
               (singular, plural) order. If no sequence is given, we'll
               take a guess base on some very coarse rules about making
               nouns plural in English. (If root is a verb, we're going
               to guess wrong, so supply your own suffix values for
               verbs.)
    fmt      - A map-based format string that produces this function's
               return value. The default format is simply "%(word)s",
               but a "count" value is also available. For example:

                   decline('rabbit',5)

               simply returns 'rabbits', but

                   decline('rabbit',5,fmt='%(count)d %(word)s')

               returns '5 rabbits'.