Rename the --show-me option of ts to --filename
Issue: Add -z option to ts command #10
This change doesn't adderess the theme of Issue #10, but renaming -s
(--show-me) to --filename is listed as an afterthought in that issue.
Because I'm about to deploy ts into a production environment that would
otherwise use --show-me, and thereby setting myself up for the
maintenance overhead of remembering update that environment once I get
this option renamed, I've just renamed the blasted option now. Technical
dept, to say nothing of the attendant cognitive load, avoided.