Add un-indenting to the "indent" command's talents
I occasionally need to unindent text, so the "indent" command will now,
when given 0 for the number of spaces to indent, remove all spaces from
the first non-blank line and then remove that same number of spaces from
the start of every subsequent line. So ...

|    rules=[
|      Suffixer('','es',
|        test=(lambda s: any([
|          s.endswith(e) for e in ('s','sh','ch','x')
|        ])),
|        text=('A word ending with "s", "sh", "ch", or "x" '
|              'becomes plural by ending "es" instead.')
|      ),
|      Suffixer('y','ies',replace=-1,
|        test=(lambda s:
|          len(s)>2 and s[-1]=='y' and s[-2] not in 'aeiou'
|        ),
|        text=('A word ending with y becomes plural by ending with y '
|              'instead, unless preceded by a vowel.')
|      ),
|      Suffixer('','s'),
|    ]

becomes ...

|rules=[
|  Suffixer('','es',
|    test=(lambda s: any([
|      s.endswith(e) for e in ('s','sh','ch','x')
|    ])),
|    text=('A word ending with "s", "sh", "ch", or "x" '
|          'becomes plural by ending "es" instead.')
|  ),
|  Suffixer('y','ies',replace=-1,
|    test=(lambda s:
|      len(s)>2 and s[-1]=='y' and s[-2] not in 'aeiou'
|    ),
|    text=('A word ending with y becomes plural by ending with y '
|          'instead, unless preceded by a vowel.')
|  ),
|  Suffixer('','s'),
|]

This is very handy when I need to paste a Python snippet into an
interactive Python session.

The behavior of the "indent" command when given a positive number of
spaces to indent, its behavior has not changed.