Add gitch() function to handy.py
getch(prompt=None, echo=False)

Read a single keystroke from stdin. The user needn't press Enter. This
function returns the character as soon has it is typed. The character is
not echoed to the screen unless the "echo" argument is True.

If "prompt" is some true value, write that string to standard output
before getting the input character, and then after the input, write a
newline character to standard output.