Add human byte-size formatting to english.py
Classes:
    Humanizer
        DecHumanizer
        DecimalHumanizer

Humanizer is the base class and handles (poorly at the moment)
expressing numbers of bytes an byte, kilobytes, megabytes, etc. so
they'll be easier to read.

For debugging purposes, I need to do some logging, and that's
problematic on a Mac because its logging is so crazy. That's why this is
getting checked in broken.