More work on english.py
I've replaced the precision parameter used in HumanBytes and its
subclasses with a formatter function for more flexability in how numeric
values are formatted. The DecimalHumanBytes and DecHumanBytes classes
are done (unless something comes up). I still need to create
BinaryHumanBytes and BinHumanBytes classes that use a divisor of 1024
rather than 1000, but that won't take long.

I've also improved debug.py a bit. DebugChannel's indent() method now
assumes an indent level increase of 1 if none is given, and there's an
undent() function that works the same way, but in reverse.