Finish HumanBytes and it subclasses
HumanBytes
    BinHumanBytes
    BinaryHumanBytes
    DecHumanBytes
    DecimalHumanBytes

We now have Binary and Decimal specializations of HumanBytes, and they
all accept arbitrary numeric formatter functions, the default being

    lambda x:pnum(x,sep=None,digits=3)

to limit significant digits to 3.

The units for BinaryHumanBytes are:

    byte
    kibibyte
    mebibyte
    gibibyte
    tebibyte
    pebibyte
    exbibyte
    zebibyte
    yobibyte

They sound goofy, so nobody actually talks that way, but those are the
official names.

The units for DecimalHumanBytes are:

    byte
    kilobyte
    megabyte
    gigabyte
    terabyte
    petabyte
    exabyte
    zettabyte
    yottabyte

The BinHumanBytes and DecHumanBytes use the same abbreviations for both
of the above sets of units:

    B
    KB
    MB
    GB
    TB
    PB
    EB
    ZB
    YB