Replace reduce's --pre option with --map
I did a bad job of naming reduce's --pre option. I named it that because
I thought of it as a "pre-processing" funciton to be applied to each
input value, replacing each value with the function's result. This, of
course, is the very definition of "map," so I've renamed --pre to --map.
This is a strictly cosmetic update. No actually functionality has been
changed in any way.