Port phonetics.py to Python 3
I've also moved the functionality of the old pa command (now removed)
into the "main" section of this module and have set up 5 symlinks to
point to it:

    nato:   NATO nearly universal at this point
    police: Police (maybe during the "Adam 12" TV show)
    raf:    RAF during WWII
    us:     US during WWII
    wu:     Western Union

usage: nato [-h] [-d D] [-i] [-l] [-L] [text ...]

This program spells words in a phonetic alphabet. For instance, in the
NATO dialect, 'a' becomes 'Alpha', b becomes 'Bravo', and so on. The
phonetic spelling of each word is written to one line of standard
output. Supported dialects are nato, police, raf, us, and wu.

positional arguments:
  text               One or more words to output phonetically.

optional arguments:
  -h, --help         show this help message and exit
  -d D, --dialect D  The phonetic dialect to be used. (default: nato)
  -i, --ignore-case  Ignore the case of the input. By default, F ==>
                     FOXTROT while f ==> Foxtrot, for example.
  -l, --list         Just list the default or given phonetic dialect.
  -L, --list-all     List words from all available dialects in tabular
                     form.