Improve english.join()
I've beefed up english.join a bit before donating it to some code I
writing for $work$. It now knows about "bracketing separators," which
open and close each item rather than simply separating them. See below
for details.

join(seq, con='and', sep=None)

Return the string values in the given sequence as an English
sequence-phrase, employing the given conjunction and seaparator when
called for by the rules of standard English. If seq is not a sequence,
it MUST be an object with a split() method, which will be used to
convert it to a sequence.

The conjunction defaults to 'and'. The separator defaults to a comma,
but see SEPERATOR DEFAULTS below.

ITEMS
  0   An empty sequence returns an empty string.

  1   A one-item sequence simply returns that item.

  2   A two-item sequence, returns a string of the form:

      "1 {con} 2"

  3+  A sequence with more than two items, returns a string of the
      form:

      "1, 2, ..., {con} N"

      The separator ALWAYS precedes the conjunction in such a case. We
      are not barbarians.

SEPARATOR DEFAULTS:
  ,   This is the first default and is most often used.

  ;   If a comma is found in any sequence item, semicolon is the next
      default.

  (   If both a comma and semicolon are found among the sequence
  {   items, each of these separators is tried in the order shown.
  [   These are bracketing separators and indicate a bracketing pair
  <   to surround each sequence item. Each one is tried until it is
  "   found that none of them, including its closing character, exists
  '   in any sequence item.

  If, weirdly, each of these potential separators is found among the
  sequence items, then the last one is used, and this function relies
  on the caller to have balanced all such groupings within each item.
  In this last case, I'm a little tempted to raise a ValueError
  containing an admonition to restructure what the caller wishes to
  express, but this isn't Grammarly.

BRACKETING SEPARATORS
You can actually force the use of (parentheses), {braces}, [square] or
<angle> brackets, or single (') or double (") quotes around each
sequence item by giving the opening character as the separator. The
join() function will understand that each item should begin with that
characterr and then end with its mate.

The first thing to know about bracketing separators is that they come
into play when there are two or more items in the sequence, as opposed
to three or mor in the case of non-bracketing separators.

  > english.join(('a and b',),sep='(')
  'a and b'

  > english.join(('a and b','c or d'),sep='(')
  '(a and b) and (c or d)'

  > english.join(('a and b','c or d','e but not f'),sep='(')
  '(a and b), (c or d), and (e but not f)'

When using quotes as the separator, we do our best to make sure the
commas precede each closing quote.

   > print(english.join(('a and b','c or d','e but not f'),sep='"'))
  "a and b," "c or d," and "e but not f"

But we're smart enough not to double up commas:

  > print(english.join(('weird,','maybe wrong','¯\_(ツ)_/¯'),'but','"'))
  "weird,", "maybe wrong," but "¯\_(ツ)_/¯"

All this can produce some odd-looking output, but it's as right as we
know how to make it:

  > print(english.join(list(',;(){}[]<>')))
  ",", ";," "(," ")," "{," "}," "[," "]," "<," and ">"

Notice that the separator defaulted to the double quote because that
was next in line after the '<' default separator, which one of the
sequence items contained.