Add some Jinja2 features to the jinja command
The --line-comment (or -C) option sets a string that, if it appears at
the start of a line, will make the rest of that line a comment.

The --line-statement (or -S) option sets a string that, if it appears at
the start of a line, will make the rest of that line a statement. For,
for instance, "-S --" might be interesting when templating SQL.

the --extension (or -x) option adds a given extension (e.g. "-x do") to
the Jinja2 environment, enabling recognition of non-Jinja2-standard tags
in the template.