Update how class Stardate is instantiated
Before, either int or float would be interpreted as a Unix epoch value.
Now, int is a Unix epoch value, but a float is interpreted as a stardate
value (1000ths of a year starting at Jan 1, 2323).

I've also made stardate.py executable and added a command line interface
to be used for testing or for working with stardate values from the
command line.