Port the slice command to Python 3
I dang-near rewrote this script because `file` is no longer something we
can subclass. So I wrote a FileReader class that reads lines, keeps
track of the line number and file position of the most recently read
line, and can be uses as both a context manager and an iterator. But
that was actually a big improvement and let me simplify the remainder of
the code considerably. (I should have done it this way to begin with.)