Add CSV.py to our handy Python modules
WARNING! This module has just barely been written and has certainly
never been tested. It will probably crash someplace if you try to use
it, but even if it doesn't crash, DO NOT TRUST IT (yet).

This CSV module is mostly copied and slightly augmented from Python's
standard csv module to accommodate reading CSV files that might have
commented lines. This is VERY non-standard, but ... since when did CSV
have a standard? In any case, the default comment character is # and
will obscure any line where it is the first character.

This CSV module also adds provision for specifying a dialect string.