Add unit tests for CSV.py
I'm kind of liking Python's unittest module (finally), and putting it
into a "MODULE-test" sidecar script makes sense for keeping CSV.py clean
and automating this testing.

I've also backed way off on the amount of specialization CSV.py
provides. It's still supports dialect specification strings, but I've
removed all the logic for supporting comments in CSV files because I was
having to do some really ugly/bad/stupid coding to get even close to the
functionality I wanted, and all for a feature that would be so rarely
used that it really ought to be implemented in the calling application
anyway. (That's sour grapes, programmer style.)