Merge cutcsv's functionality into the csv command
I've added --join and --keep (sort of the converse of "cut," I suppose)
to the csv command, so cutcsv is no longer necessary. I've also moved
cutcsv into the "attic" directory.

This new version of csv also includes a bit of logic so it will find our
"handy" modules even if the shell's PYTHONPATH doesn't point to them, so
long as the csv command is either installed in the usual directory
structure or is being run from the local git repository.