Fix bug in PYTHONPATH composition logic
The recent addition of code to the top of csv that makes sure "handy's"
python modules are available was buggy (raising an error when everything
was fine to begin with). This has been corrected, an I've received a
sound trashing for failing to see the very simple problem with the code
I was writing.