Add --file and --file-last to progress script
The --file command line option changes the meanings if the pos and limit
command line arguments. pos is interpreted as a regular expression, and
limit is a filename. The position of the first occurrance of the pos RE
within the file is used as the pos value, and the number of bytes in the
file is used as the limit value. In a world that includes Unicode and
different line-ending character sequences, the implementation I've used
is imperfect. But I don't think you'll notice unless the file's use of
Unicode is particularly heavy or if the file is quite large.

The new --file-last option is just like --file, but the position of the
last occurrance of the RE is used.

I've also added --format as a synonym for --fmt. Spell this option
either way. While testing, I also removed the restriction against using
--fmt with --percent. This has the effect of removing --percent's
default formatting of "(%d%%)", in favor of the global default "(%d)",
so if you want the percent sign back you'll have to put it there
yourself using --fmt when you use --percent. ¯\_(ツ)_/¯