mark handles bad Unicode data better now
One of my main uses for the mark script is to colorize Unix logs, e.g.
/var/log/message, as that content goes flying by. And it seems syslog
(or something) will, once in a blue moon, insert some random byte (not
even a character) into its output stream.

So I've added some UnicodeDecodeError exception handling that writes the
offending line to standard error and then moves on to the next line
without any further processing of the line containing the bad data.