Improve how mark handles unreadable Unicode
mark's get_input_lines() function was getting really messy, and I needed
to improve the error message it writes to standard error when it reads
bad Unicode from its input stream. So I've farily radically restructured
this generator function's logic, and it's much more streamlined now.

Unfortunately, when Python finds undecipherable Unicode in an input
stream's buffer, the whole buffer is lost. The next read, so far as I
can tell, triggers a new buffer to be read from the input stream, which
means it will come from some operationally random location later in that
stream, not necessarily (and most likely not) the beginning of a line,
or even the start of a valid Unicode character.

Given all this, I've expanded the error message printed to standard
error in this situcation to communicate this, albeit tersely, to the
user.