Fixed install.py's problem with relative symlinks
I'd gotten the my wires crossed in my head when I wrote the previous
version of the File class's link() and __call__() methods. After
thinking through it more clearly, it was a simple rewrite of the
algorithm (which turned out to be simpler). I'm still testing it
because I think there are some cases involving forcing new symlinks to
be created that aren't working as advertised. (See TODO comments
install.py.)