Add OSError handling to install.py
The File and Folder classes now handle OSError exceptions gracefully.
They store (in the instance's "exception" attribute and in Target's
class "exceptions" attribute, which is a list) and announce (on
standard) the errors and continue so that other files can be installed.
This gives calling code the option of checking for exceptions after each
operation, after a group of operations, or both.

The install script will now exit with a non-zero return code if any
OSError exceptions occurred during execution.

I've also dumbed down this code to run under Python 3.6. RHEL runs way,
WAY behind the curve on Python versions.