Move some old code into the attic
It's weird to me at this point that I write datecycle in C. It would be
much more portable if implemented in Python, and logfile rotation does
exactly need to be speedy. (I don't know what I was thinking, but this
is very old code.)

The numlines command just counts newlines in a file. I have no idea why
I wrote that or thought it worthy of being this repo. I'll likely remove
it entirely soon.