Add --lambda option to csv script
There's a new --lambda option for csv that lets the user give a lambda
expression (minus the "lambda" keyword) accepting two arguments, row
number (starting with 1) and row (as a list). The function must return
either the row, possibly modified, or None, in which case the current
row is discarded entirely, and processing continues with the new row.

For example:

    --lambda "i,r:r[1:3] if r[0]=='Arizona' else None"

The above lambda function's arguments are i (the record number starting
with 1) and r (a list containing the values for the given row). This
particular lambda function doesn't use i, but it keeps only columns 2
and 3 if column 1 is "Arizona". Otherwise, it skips that input row.

I also fixed some minor bugs having to do with handling numeric values.