Add --text output format to qr script
I've added a smaller version of the "terminal" output formap called
"text." The terminal output format uses "██" (two characters) for
pixels, but the text output format is a quarter of the size. It uses one
of the block characters ("▀", "▄", and "█") to print two rows of pixels
at once. This is now the default output format.